+++ src/commands/webhooks.ts
@@ -1,0 +1,6 @@
+import * as vscode from 'vscode';
+import { IContext, showErrorMessage } from '../common';
+
+export async function createWebhook(context: IContext) {
+    // TODO
+}